requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-stdimage==4.1.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.2
Pillow==5.4.1
progressbar2==3.42.0
python-utils==2.3.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
stripe==2.24.0
tablib==0.13.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0

psycopg2
gunicorn
dj-database-url
whitenoise== 3.3.1